Add possibility to transform BufferedImage to Alto4
