+++ jochre_core/src/main/java/com/joliciel/jochre/doc/ImageDocumentExtractor.java
@@ -18,22 +18,18 @@
 //////////////////////////////////////////////////////////////////////////////
 package com.joliciel.jochre.doc;
 
-import java.awt.image.BufferedImage;
-import java.io.File;
-import java.io.FilenameFilter;
-
-import javax.imageio.ImageIO;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import com.joliciel.talismane.utils.Monitorable;
 import com.joliciel.talismane.utils.MultiTaskProgressMonitor;
 import com.joliciel.talismane.utils.ProgressMonitor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.imageio.ImageIO;
+import java.awt.image.BufferedImage;
+import java.io.InputStream;
 
 /**
- * An interface for extracting a JochreDocument from an image File (jpeg, gif or
- * png).
+ * An interface for extracting a JochreDocument from an image input stream.
  * 
  * @author Assaf Urieli
  *
@@ -42,12 +38,14 @@ public class ImageDocumentExtractor implements Monitorable, Runnable {
   private static final Logger LOG = LoggerFactory.getLogger(ImageDocumentExtractor.class);
   private final SourceFileProcessor documentProcessor;
   private MultiTaskProgressMonitor currentMonitor;
-  private final File imageFile;
+  private final BufferedImage image;
+  private final String fileName;
   private int pageNumber = 1;
 
-  public ImageDocumentExtractor(File imageFile, SourceFileProcessor documentProcessor) {
+  public ImageDocumentExtractor(BufferedImage image, String fileName, SourceFileProcessor documentProcessor) {
     this.documentProcessor = documentProcessor;
-    this.imageFile = imageFile;
+    this.image = image;
+    this.fileName = fileName;
   }
 
   @Override
@@ -56,46 +54,31 @@ public class ImageDocumentExtractor implements Monitorable, Runnable {
   }
 
   public JochreDocument extractDocument() {
-    LOG.debug("ImageDocumentExtractorImpl.extractDocument");
+    LOG.debug("InputStreamDocumentExtractor.extractDocument");
     try {
-      File[] files = new File[1];
 
-      if (imageFile.isDirectory()) {
-        files = imageFile.listFiles(new FilenameFilter() {
-
-          @Override
-          public boolean accept(File dir, String name) {
-            return (name.toLowerCase().endsWith(".png") || name.toLowerCase().endsWith(".jpg") || name.toLowerCase().endsWith(".jpeg")
-                || name.toLowerCase().endsWith(".gif") || name.toLowerCase().endsWith(".tif") || name.toLowerCase().endsWith(".tiff"));
-          }
-        });
-      } else {
-        files[0] = imageFile;
-      }
 
       JochreDocument doc = this.documentProcessor.onDocumentStart();
-      doc.setTotalPageCount(files.length);
+      doc.setTotalPageCount(1);
 
       int currentPageNumber = this.pageNumber;
-      for (File file : files) {
-        JochrePage page = this.documentProcessor.onPageStart(currentPageNumber++);
 
-        BufferedImage image = ImageIO.read(file);
-        String imageName = file.getName();
+      JochrePage page = this.documentProcessor.onPageStart(currentPageNumber++);
 
-        if (currentMonitor != null && documentProcessor instanceof Monitorable) {
-          ProgressMonitor monitor = ((Monitorable) documentProcessor).monitorTask();
-          double percentAllotted = (1 / (double) (files.length));
-          currentMonitor.startTask(monitor, percentAllotted);
-        }
+      String imageName = this.fileName;
 
-        documentProcessor.onImageFound(page, image, imageName, 0);
-        if (currentMonitor != null && documentProcessor instanceof Monitorable) {
-          currentMonitor.endTask();
-        }
-
-        this.documentProcessor.onPageComplete(page);
+      if (currentMonitor != null && documentProcessor instanceof Monitorable) {
+        ProgressMonitor monitor = ((Monitorable) documentProcessor).monitorTask();
+        currentMonitor.startTask(monitor, 1.0);
       }
+
+      documentProcessor.onImageFound(page, image, imageName, 0);
+      if (currentMonitor != null && documentProcessor instanceof Monitorable) {
+        currentMonitor.endTask();
+      }
+
+      this.documentProcessor.onPageComplete(page);
+
       this.documentProcessor.onDocumentComplete(doc);
       this.documentProcessor.onAnalysisComplete();
 
@@ -109,7 +92,7 @@ public class ImageDocumentExtractor implements Monitorable, Runnable {
       LOG.error("Exception while processing document", e);
       throw new RuntimeException(e);
     } finally {
-      LOG.debug("Exit ImageDocumentExtractorImpl.extractDocument");
+      LOG.debug("Exit InputStreamDocumentExtractor.extractDocument");
     }
   }
