+++ jochre_core/src/main/java/com/joliciel/jochre/doc/ImageFileDocumentExtractor.java
@@ -18,36 +18,36 @@
 //////////////////////////////////////////////////////////////////////////////
 package com.joliciel.jochre.doc;
 
-import com.joliciel.talismane.utils.Monitorable;
-import com.joliciel.talismane.utils.MultiTaskProgressMonitor;
-import com.joliciel.talismane.utils.ProgressMonitor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.FilenameFilter;
-import java.io.InputStream;
+
+import javax.imageio.ImageIO;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.joliciel.talismane.utils.Monitorable;
+import com.joliciel.talismane.utils.MultiTaskProgressMonitor;
+import com.joliciel.talismane.utils.ProgressMonitor;
 
 /**
- * An interface for extracting a JochreDocument from an image input stream.
+ * An interface for extracting a JochreDocument from an image File (jpeg, gif or
+ * png).
  * 
  * @author Assaf Urieli
  *
  */
-public class InputStreamDocumentExtractor implements Monitorable, Runnable {
-  private static final Logger LOG = LoggerFactory.getLogger(InputStreamDocumentExtractor.class);
+public class ImageFileDocumentExtractor implements Monitorable, Runnable {
+  private static final Logger LOG = LoggerFactory.getLogger(ImageFileDocumentExtractor.class);
   private final SourceFileProcessor documentProcessor;
   private MultiTaskProgressMonitor currentMonitor;
-  private final InputStream imageInputStream;
-  private final String fileName;
+  private final File imageFile;
   private int pageNumber = 1;
 
-  public InputStreamDocumentExtractor(InputStream imageInputStream, String fileName, SourceFileProcessor documentProcessor) {
+  public ImageFileDocumentExtractor(File imageFile, SourceFileProcessor documentProcessor) {
     this.documentProcessor = documentProcessor;
-    this.imageInputStream = imageInputStream;
-    this.fileName = fileName;
+    this.imageFile = imageFile;
   }
 
   @Override
@@ -56,32 +56,46 @@ public class InputStreamDocumentExtractor implements Monitorable, Runnable {
   }
 
   public JochreDocument extractDocument() {
-    LOG.debug("InputStreamDocumentExtractor.extractDocument");
+    LOG.debug("ImageDocumentExtractorImpl.extractDocument");
     try {
+      File[] files = new File[1];
 
+      if (imageFile.isDirectory()) {
+        files = imageFile.listFiles(new FilenameFilter() {
+
+          @Override
+          public boolean accept(File dir, String name) {
+            return (name.toLowerCase().endsWith(".png") || name.toLowerCase().endsWith(".jpg") || name.toLowerCase().endsWith(".jpeg")
+                || name.toLowerCase().endsWith(".gif") || name.toLowerCase().endsWith(".tif") || name.toLowerCase().endsWith(".tiff"));
+          }
+        });
+      } else {
+        files[0] = imageFile;
+      }
 
       JochreDocument doc = this.documentProcessor.onDocumentStart();
-      doc.setTotalPageCount(1);
+      doc.setTotalPageCount(files.length);
 
       int currentPageNumber = this.pageNumber;
+      for (File file : files) {
+        JochrePage page = this.documentProcessor.onPageStart(currentPageNumber++);
 
-      JochrePage page = this.documentProcessor.onPageStart(currentPageNumber++);
+        BufferedImage image = ImageIO.read(file);
+        String imageName = file.getName();
 
-      BufferedImage image = ImageIO.read(this.imageInputStream);
-      String imageName = this.fileName;
+        if (currentMonitor != null && documentProcessor instanceof Monitorable) {
+          ProgressMonitor monitor = ((Monitorable) documentProcessor).monitorTask();
+          double percentAllotted = (1 / (double) (files.length));
+          currentMonitor.startTask(monitor, percentAllotted);
+        }
 
-      if (currentMonitor != null && documentProcessor instanceof Monitorable) {
-        ProgressMonitor monitor = ((Monitorable) documentProcessor).monitorTask();
-        currentMonitor.startTask(monitor, 1.0);
+        documentProcessor.onImageFound(page, image, imageName, 0);
+        if (currentMonitor != null && documentProcessor instanceof Monitorable) {
+          currentMonitor.endTask();
+        }
+
+        this.documentProcessor.onPageComplete(page);
       }
-
-      documentProcessor.onImageFound(page, image, imageName, 0);
-      if (currentMonitor != null && documentProcessor instanceof Monitorable) {
-        currentMonitor.endTask();
-      }
-
-      this.documentProcessor.onPageComplete(page);
-
       this.documentProcessor.onDocumentComplete(doc);
       this.documentProcessor.onAnalysisComplete();
 
@@ -95,7 +109,7 @@ public class InputStreamDocumentExtractor implements Monitorable, Runnable {
       LOG.error("Exception while processing document", e);
       throw new RuntimeException(e);
     } finally {
-      LOG.debug("Exit InputStreamDocumentExtractor.extractDocument");
+      LOG.debug("Exit ImageDocumentExtractorImpl.extractDocument");
     }
   }
